I refactored the monolith app.py into an application factory utilizing blueprints and dynamic environment settings
